refactor(redis): split catalog and value boundaries
